Adds reply option to comments

+++ api/routers/comment.py
@@ -3,7 +3,7 @@ from typing import Annotated
 from fastapi import Depends, status
 from sqlmodel import select
 
-from api.error import NotFound, ErrorResponses
+from api.error import ConflictError, NotFound, ErrorResponses
 from api.models import SessionDep, audit
 from api.models.db import Comment
 from api.models.insert import Insert
@@ -117,10 +117,19 @@ async def deleteComment(
     commentID: int,
     session: SessionDep,
     userId: Annotated[str, Depends(getUserId)],
+    force: bool = False,
 ) -> None:
     comment = session.get(Comment, commentID)
     if not comment:
         raise NotFound(detail="Comment not found")
+    if len(comment.children) > 0 and force is False:
+        raise ConflictError(
+            detail=[
+                f"Comment has {len(comment.children)} {'child' if len(comment.children) == 1 else 'children' }.",
+                "Use force and cascade (?force=true) to delete anyway.",
+                "This will also delete the extras.",
+            ]
+        )
     session.delete(comment)
     session.commit()
     audit(session, 2, comment, userId)
